Add conftest in tests
