+++ backend/tests/github_api_test.py
@@ -1,0 +1,1 @@
+import pytest
